02_airflow/requirements.txt
pyarrow
apache-airflow-providers-google
pandas
requests
beautifulsoup4